Game server: removed CKingSystem::m_byTerritoryTariff, the tariff is now zone-specific.

+++ src/Ebenezer/User.cpp
@@ -827,7 +827,7 @@ void CUser::SetZoneAbilityChange()
 	result	<< GetMap()->canTradeWithOtherNation()
 			<< GetMap()->GetZoneType()
 			<< GetMap()->canTalkToOtherNation()
-			<< GetMap()->GetTariff();
+			<< uint16(GetMap()->GetTariff());
 
 	Send(&result);
 }
